+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.83 (Pending):
 * Add helper class `OpenMQUtil` to help test OpenMQ state during tests.
+* Add several methods to `OpenMQContainer` that support creating queues and topics
+  within the broker.
 
 ## 0.82:
 * Add helper methods `AbstractServerTest.ctran` that clear the EntityManager before running block.